Move Preview To Child Component
- The parent component now focuses solely on managing the overall upload process, while the child component handles the rendering and logic for displaying image previews.
